[wip] implements new multi-file parquet iterator, moves all iterators to separate file, and adds some unit tests
