+++ .gitignore
@@ -45,3 +45,4 @@ snapshot.json
 
 # Offline Tests
 provider/generic_records_*.parquet
+provider/part-000*.parquet
